+++ src/sll/map.c
@@ -9,6 +9,7 @@
 #include <sll/runtime_object.h>
 #include <sll/static_object.h>
 #include <sll/types.h>
+#include <sll/util.h>