+++ src/sll/operator.c
@@ -13,6 +13,7 @@
 #include <sll/static_object.h>
 #include <sll/string.h>
 #include <sll/types.h>
+#include <sll/util.h>
 #include <math.h>